rag_chatbot_api/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
langchain
langchain-community
langchain-openai
langchain-google-genai
tiktoken
psycopg2-binary
sqlalchemy
asyncpg
markdown
beautifulsoup4
unstructured